+++ libs/data-mapper-v2/src/components/addSchema/AddSchemaPanel.tsx
@@ -1,13 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { getSelectedSchema } from '../../core';
+import { DataMapperFileService, getSelectedSchema } from '../../core';
 import { setInitialSchema } from '../../core/state/DataMapSlice';
 import { closePanel, ConfigPanelView, openDefaultConfigPanelView } from '../../core/state/PanelSlice';
 import type { AppDispatch, RootState } from '../../core/state/Store';
 import { LogCategory, LogService } from '../../utils/Logging.Utils';
 import { convertSchemaToSchemaExtended, getFileNameAndPath } from '../../utils/Schema.Utils';
-import type { SchemaFile } from './AddOrUpdateSchemaView';
-import { AddOrUpdateSchemaView, UploadSchemaTypes } from './AddOrUpdateSchemaView';
-//import { DefaultConfigView } from './DefaultConfigView';
+import { AddOrUpdateSchemaView, UploadSchemaTypes, type SchemaFile } from './AddOrUpdateSchemaView';
 import { DefaultButton, PrimaryButton } from '@fluentui/react';
 import type { DataMapSchema } from '@microsoft/logic-apps-shared';
 import { SchemaType } from '@microsoft/logic-apps-shared';
@@ -25,14 +23,14 @@ const schemaFileQuerySettings = {
 
 export interface ConfigPanelProps {
   onSubmitSchemaFileSelection: (schemaFile: SchemaFile) => void;
-  readCurrentSchemaOptions?: () => void;
   schemaType: SchemaType;
 }
 
-export const AddSchemaDrawer = ({ readCurrentSchemaOptions, onSubmitSchemaFileSelection, schemaType }: ConfigPanelProps) => {
+export const AddSchemaDrawer = ({ onSubmitSchemaFileSelection, schemaType }: ConfigPanelProps) => {
   const dispatch = useDispatch<AppDispatch>();
   const intl = useIntl();
   const styles = useStyles();
+  const fileService = DataMapperFileService();
 
   const curDataMapOperation = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation);
   const currentPanelView = useSelector((state: RootState) => {
@@ -174,10 +172,10 @@ export const AddSchemaDrawer = ({ readCurrentSchemaOptions, onSubmitSchemaFileSe
 
   // Read current schema file options if method exists
   useEffect(() => {
-    if (readCurrentSchemaOptions) {
-      readCurrentSchemaOptions();
+    if (fileService && fileService.readCurrentSchemaOptions) {
+      fileService.readCurrentSchemaOptions();
     }
-  }, [readCurrentSchemaOptions]);
+  }, [fileService]);
 
   const onRenderFooterContent = useCallback(() => {
     if (currentPanelView === ConfigPanelView.DefaultConfig) {
@@ -225,14 +223,7 @@ export const AddSchemaDrawer = ({ readCurrentSchemaOptions, onSubmitSchemaFileSe
 
   return (
     <div className={styles.root}>
-      <InlineDrawer
-        open={!!currentPanelView}
-        size="small"
-        className={mergeClasses(
-          selectedSchemaFile ? styles.fileSelectedDrawer : styles.drawer,
-          schemaType === SchemaType.Source ? styles.leftDrawer : styles.rightDrawer
-        )}
-      >
+      <InlineDrawer open={!!currentPanelView} size="small" className={mergeClasses(styles.drawerWrapper)}>
         <AddOrUpdateSchemaView
           schemaType={schemaType}
           selectedSchema={selectedSchema?.name}