+++ libs/data-mapper-v2/src/core/DataMapperDesignerContext.tsx
@@ -1,7 +1,8 @@
-import { createContext } from 'react';
+import { createContext, type MutableRefObject } from 'react';
 
 export interface DataMapperDesignerContext {
   readOnly?: boolean;
+  canvasRef?: MutableRefObject<HTMLDivElement | null>;
 }
 
 export const DataMapperWrappedContext = createContext<DataMapperDesignerContext | null>(null);