+++ CHANGELOG.md
@@ -1,4 +1,95 @@
 # Logic Apps Designer
+## [4.27.0](https://github.com/Azure/LogicAppsUX/compare/v4.26.0...v4.27.0) (2024-06-08)
+
+
+### Features
+
+* **vscode:** Introduce .NET 8 to custom code and extension dependencies ([#4947](https://github.com/Azure/LogicAppsUX/issues/4947)) ([4e14837](https://github.com/Azure/LogicAppsUX/commit/4e14837ab28b5f07590e75bd540a7d2877dad4c7))
+
+
+### Bug Fixes
+
+* **designer:** Fix Real API tests ([#4942](https://github.com/Azure/LogicAppsUX/issues/4942)) ([41ca069](https://github.com/Azure/LogicAppsUX/commit/41ca06946d317ab804c56158210e301219879257))
+* **designer:** Stringifying appsetting values during connection creation ([#4938](https://github.com/Azure/LogicAppsUX/issues/4938)) ([3ed99b9](https://github.com/Azure/LogicAppsUX/commit/3ed99b9d0937280d2f8eb2848424f4319cec56ec))
+* **vscode:** Update Workflows Webjobs sdk version ([#4952](https://github.com/Azure/LogicAppsUX/issues/4952)) ([187321e](https://github.com/Azure/LogicAppsUX/commit/187321edb9952918314e7ca415dc26cc52a00066))
+
+## [4.26.0](https://github.com/Azure/LogicAppsUX/compare/v4.25.0...v4.26.0) (2024-06-06)
+
+## [4.25.0](https://github.com/Azure/LogicAppsUX/compare/v4.24.0...v4.25.0) (2024-06-05)
+
+
+### Features
+
+* **designer-ui:** Update file picker to use Fluent UI v9 components ([#4917](https://github.com/Azure/LogicAppsUX/issues/4917)) ([cdfa7ae](https://github.com/Azure/LogicAppsUX/commit/cdfa7aed25dc07f856de88c5fc609259a3b1622d))
+* **vscode:** Introduce creation of rules code project  ([#4920](https://github.com/Azure/LogicAppsUX/issues/4920)) ([916c529](https://github.com/Azure/LogicAppsUX/commit/916c5290f4b4c4c002fbe298fa0a53758ba76b26))
+
+
+### Bug Fixes
+
+* **designer:** Do not show label for array editor in collapsed view ([#4932](https://github.com/Azure/LogicAppsUX/issues/4932)) ([2925b87](https://github.com/Azure/LogicAppsUX/commit/2925b87efddf34ff745b22922b71ac8f2285c784))
+
+## [4.24.0](https://github.com/Azure/LogicAppsUX/compare/v4.23.0...v4.24.0) (2024-06-03)
+
+
+### Features
+
+* **designer:** Adding support for dynamic input schema in dynamic parameters ([#4844](https://github.com/Azure/LogicAppsUX/issues/4844)) ([caa69b6](https://github.com/Azure/LogicAppsUX/commit/caa69b62489df9aba149306134ae920ef5da272c))
+
+## [4.23.0](https://github.com/Azure/LogicAppsUX/compare/v4.22.0...v4.23.0) (2024-05-31)
+
+
+### Features
+
+* **Data Mapper:** Get schema details from services, removing mock ([#4923](https://github.com/Azure/LogicAppsUX/issues/4923)) ([b09b975](https://github.com/Azure/LogicAppsUX/commit/b09b975d1a7a125f4fb8262ec4d2e6515d1f9cce))
+* **Data Mapper:** Moving away from mock schema list ([#4912](https://github.com/Azure/LogicAppsUX/issues/4912)) ([789043b](https://github.com/Azure/LogicAppsUX/commit/789043bc4daceea39b3767d603bd00f4848bffa8))
+* **designer:** Add ability for host to send in overrides for internationalized strings ([#4907](https://github.com/Azure/LogicAppsUX/issues/4907)) ([b6be06a](https://github.com/Azure/LogicAppsUX/commit/b6be06a85e093f5eb655d6bbf899bfb352454918))
+
+
+### Bug Fixes
+
+* **designer:** Turn off spell check in editors ([#4902](https://github.com/Azure/LogicAppsUX/issues/4902)) ([6332975](https://github.com/Azure/LogicAppsUX/commit/6332975dedca796f07327e317f64aa5eca7250bf))
+* **designer:** Using custom Label with ariaLabel for parameter editors for a11y ([#4889](https://github.com/Azure/LogicAppsUX/issues/4889)) ([2778673](https://github.com/Azure/LogicAppsUX/commit/27786736af369feee8992abc19943311a2ecd04c))
+* **designer:** Using UI custom Label everywhere in the designer to be consistent for a11y ([#4903](https://github.com/Azure/LogicAppsUX/issues/4903)) ([e8abdc5](https://github.com/Azure/LogicAppsUX/commit/e8abdc597da445a711a7de2e100113410540acff))
+* **meta:** Update pnpm version in coverage-report ([#4921](https://github.com/Azure/LogicAppsUX/issues/4921)) ([7011da0](https://github.com/Azure/LogicAppsUX/commit/7011da065188ffba897d8228c1b9bd4892d437f3))
+* **meta:** Update pnpm version to use ([#4919](https://github.com/Azure/LogicAppsUX/issues/4919)) ([4d3df02](https://github.com/Azure/LogicAppsUX/commit/4d3df02f6df97971d7934a227dee47f5932b1726))
+
+## [4.22.0](https://github.com/Azure/LogicAppsUX/compare/v4.21.0...v4.22.0) (2024-05-28)
+
+
+### Bug Fixes
+
+* **designer-ui:** Allow color picker sliders to be accessed by keyboard ([#4884](https://github.com/Azure/LogicAppsUX/issues/4884)) ([98fb4bc](https://github.com/Azure/LogicAppsUX/commit/98fb4bceec324d5b5e119844fe3f79758bfcf417))
+* **designer:** `updateExistingInputTokenTitles` is not invoked when dynamic outputs load ([#4880](https://github.com/Azure/LogicAppsUX/issues/4880)) ([8704aad](https://github.com/Azure/LogicAppsUX/commit/8704aad9d7c727c8f8eafc08817c7412f1ea7ec5))
+* **Designer:** Fixed search locale not respecting portal language settings ([#4898](https://github.com/Azure/LogicAppsUX/issues/4898)) ([65fe629](https://github.com/Azure/LogicAppsUX/commit/65fe62962588d0e55d8d64d167d81fc584fac0a4))
+
+## [4.21.0](https://github.com/Azure/LogicAppsUX/compare/v4.20.0...v4.21.0) (2024-05-23)
+
+
+### Features
+
+* **Data Mapper:** Refactoring deserialization for v2 ([#4800](https://github.com/Azure/LogicAppsUX/issues/4800)) ([395385b](https://github.com/Azure/LogicAppsUX/commit/395385b8b2b031208ed6c07ebf552e6c39ffe2d4))
+
+## [4.20.0](https://github.com/Azure/LogicAppsUX/compare/v4.19.0...v4.20.0) (2024-05-21)
+
+
+### Bug Fixes
+
+* **designer-ui:** Prevent padding on HTML toolbar from creating scrollbar ([#4870](https://github.com/Azure/LogicAppsUX/issues/4870)) ([f095b42](https://github.com/Azure/LogicAppsUX/commit/f095b42402d72fa2328da91159f2d2e7de820a72))
+
+## [4.19.0](https://github.com/Azure/LogicAppsUX/compare/v4.18.0...v4.19.0) (2024-05-20)
+
+
+### Features
+
+* **Data Mapper V2:** Creating initial Nodes/Edges with connection ([#4865](https://github.com/Azure/LogicAppsUX/issues/4865)) ([1653c0b](https://github.com/Azure/LogicAppsUX/commit/1653c0b96c81d9d81d29330a12575578011bc78f))
+* **vscode:** Show local time and UTC in overview ([#4855](https://github.com/Azure/LogicAppsUX/issues/4855)) ([3cc8511](https://github.com/Azure/LogicAppsUX/commit/3cc851188130d04da789093ce7417d8aa61e15f1))
+
+
+### Bug Fixes
+
+* **Designer:** More customcode boilerplate code... ([#4858](https://github.com/Azure/LogicAppsUX/issues/4858)) ([bc26405](https://github.com/Azure/LogicAppsUX/commit/bc264055458b00c5605f4287732d51ca85dc61d2))
+* **vscode:** Update command to build all projects for extension ([#4852](https://github.com/Azure/LogicAppsUX/issues/4852)) ([797a4e2](https://github.com/Azure/LogicAppsUX/commit/797a4e2ad4ea44748545906f96b2708df241a741))
+
 ## [4.18.0](https://github.com/Azure/LogicAppsUX/compare/v4.17.0...v4.18.0) (2024-05-16)